+++ src/main/java/ru/hukola/services/service/OrderService.java
@@ -1,14 +1,13 @@
 package ru.hukola.services.service;
 
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.web.FilterChainProxy;
 import org.springframework.stereotype.Service;
 import ru.hukola.services.model.Order;
 import ru.hukola.services.repository.OrderRepository;
+import ru.hukola.services.model.User;
 
 import java.util.Collection;
 import java.util.UUID;
@@ -20,35 +19,40 @@ import java.util.UUID;
 @AllArgsConstructor
 public class OrderService {
     private OrderRepository orderRepository;
-    private ClientService clientService;
-
-    private FilterChainProxy proxy;
+    private UserService userService;
 
     public Collection<Order> findAll(int offset, int size) {
-        Page<Order> page = orderRepository.findAll(PageRequest.of(offset, size, Sort.by("date")));
+        User user = userService.getSecurityUser();
+        Page<Order> page = orderRepository.findAllByUser(user, PageRequest.of(offset, size, Sort.by("date")));
         return page.stream().toList();
     }
 
     public Order findById(UUID uuid) {
-        return orderRepository.findById(uuid).orElseThrow();
+        User user = userService.getSecurityUser();
+        return orderRepository.findByUuidAndUser(uuid, user).orElseThrow();
     }
 
     public Order save(Order order) {
+        User user = userService.getSecurityUser();
+        order.setUser(user);
         return orderRepository.save(order);
     }
 
     public Order create(UUID uuid, Order order) {
+        User user = userService.getSecurityUser();
         Order savedOrder = orderRepository.findById(uuid).orElseThrow();
         savedOrder.setAmount(order.getAmount());
         savedOrder.setPaid(order.isPaid());
         savedOrder.setClient(order.getClient());
         savedOrder.setDescription(order.getDescription());
         savedOrder.setDate(order.getDate());
+        savedOrder.setUser(user);
         orderRepository.save(savedOrder);
         return savedOrder;
     }
 
     public void delete(UUID uuid) {
-        orderRepository.deleteById(uuid);
+        if (findById(uuid) != null)
+            orderRepository.deleteById(uuid);
     }
 }
